pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wnc-proposal-assistant"
version = "0.1.0"
description = "PoC: AI-assisted donor email & grant narrative generator with lightweight RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "streamlit>=1.33.0",
    "python-docx>=1.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "ui*", "data*", "eval*", "vector*", "scripts*"]

[tool.setuptools]
include-package-data = true